+++ CHANGELOG.md
@@ -7,6 +7,7 @@ HEAD
 * Fix occasional missed ticks on timer-based implementations
 * Add `ContinuousStepper_Tone` which uses `tone()` instead of `digitalWrite()` for the step pin
 * Add `ContinuousStepper_AWF` which uses `analogWriteFrequency()` on Teensy 3 and 4
+* Add `ContinuousStepper_KhoiH` which uses Khoi Hoang's PWM libraries
 
 2.1.0
 -----